Bug: 로그인시 FireBaseError 수정
