+++ src/store/company/useCompanyStore.js
@@ -1,22 +1,31 @@
 import { create } from "zustand";
 import zukeeper from "zukeeper";
 import { v4 } from "uuid";
+import { persist, createJSONStorage } from "zustand/middleware";
 
 const useCompanyStore = create(
-  zukeeper((set) => ({
-    companyData: {
-      id: v4(),
-      name: "",
-      code: 111111,
-      address: "",
-      roles: [],
+  // zukeeper(
+  persist(
+    (set) => ({
+      companyData: {
+        id: v4(),
+        name: "",
+        code: 111111,
+        address: "",
+        roles: [],
+      },
+      setCompanyData: (companyData) => {
+        set({ companyData });
+      },
+    }),
+    {
+      name: "company",
+      storage: createJSONStorage(() => sessionStorage),
     },
-    setCompanyData: (companyData) => {
-      set({ companyData });
-    },
-  })),
+  ),
+  // ),
 );
 
-window.store = useCompanyStore;
+// window.store = useCompanyStore;
 
 export default useCompanyStore;
